zio-test: Remove the `"... + x more"` feedback from zio-test outputs (#9900)
* [zio-test] Remove the `"... + x more"` feedback from zio-test outputs

It's extremely frustrating to have a test framework which doesn't show the data when there's a diff between the result of the test execution and the expected result

I always have to add some debug `println` and relaunch my tests, to be able to see this diff. That shouldn't be the case.
zio-test should show me the result data and the expected data in its output so I can fix my test